tests/py-requirements.txt
pyyaml==3.12
flex_version>=1.2.3
verminator>=1.2.0